+++ KaisaKaavio/Properties/Resources.Designer.cs
@@ -843,6 +843,16 @@ namespace KaisaKaavio.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap Tulosta {
+            get {
+                object obj = ResourceManager.GetObject("Tulosta", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap Tyhja128 {
             get {
                 object obj = ResourceManager.GetObject("Tyhja128", resourceCulture);